Store `RepeatMetaData` into `date` field for task (#271)
It is a design mistake to use a separate field to store repeat metadata.

Reasons:
- It doesn't make a lot of sense to also keep `date` field for RepeatingTask. Before this change, we only use that to store time info, which is hard to statically ensure that we correctly populate it.
- It gives us a false sense of security that we have correctly dealt with the complexity of repeating task. Since we kept `date` field, TypeScript won't complain in a lot of places where we forget to consider that `date` in RepeatingTask has special meanings.

Therefore, I decide to make `date` to have different types for OneTimeTask and RepeatingTask, so we can use TypeScript to tell us where we did it wrong.
The majority of this commit is simply changing `repeatingPattern` to `date`, but I do discover a bug where the current behavior is wrong. In the focus view, we still always use the `date` field to decide the date when the task is due, which doesn't make sense for repeating task. I decide to ban forking or deleting ONE INSTANCE OF repeating task in focus view because it's not clear to the user what's the actual due date of a task in the focus view. We should probably add the ability back in the future and default to the earliest due date for repeating task in focus view, but it should be the concern of another pull request.

+++ frontend/src/store/reducers.ts
@@ -58,11 +58,13 @@ function patchTasks(state: State, { created, edited, deleted }: PatchTasks): Sta
       }
       const key = t.date.toDateString();
       const oldTask = state.tasks.get(t.id) || error();
-      const oldKey = oldTask.date.toDateString();
-      if (oldKey !== key) {
-        // remove first
-        const oldBucket = m.get(oldKey) || error('impossible!');
-        m.set(oldKey, oldBucket.remove(t.id));
+      if (oldTask.type === 'ONE_TIME') {
+        const oldKey = oldTask.date.toDateString();
+        if (oldKey !== key) {
+          // remove first
+          const oldBucket = m.get(oldKey) || error('impossible!');
+          m.set(oldKey, oldBucket.remove(t.id));
+        }
       }
       const set = m.get(key);
       if (set == null) {
@@ -76,10 +78,12 @@ function patchTasks(state: State, { created, edited, deleted }: PatchTasks): Sta
       if (oldTask == null) {
         return;
       }
-      const key = oldTask.date.toDateString();
-      const set = m.get(key);
-      if (set != null) {
-        m.set(key, set.remove(id));
+      if (oldTask.type === 'ONE_TIME') {
+        const key = oldTask.date.toDateString();
+        const set = m.get(key);
+        if (set != null) {
+          m.set(key, set.remove(id));
+        }
       }
     });
   });
